+++ apps/web/src/components/forms/multi-step/schemas/index.ts
@@ -3,15 +3,14 @@
 // Export combined types
 export * from '../types'
 
+// Import sustainability initiatives from config
+import { sustainabilityInitiativesStepConfig } from '@/components/forms/multi-step/steps/step-3-sustainability-initiatives/schema'
 // Import company info from new location
 import { companyInfoStepConfig } from '../steps/step-1-company-info/schema'
 // Import business model from new location
 import { businessModelStepConfig } from '../steps/step-2-business-model/schema'
 // Import sustainability practices from new location
 import { sustainabilityStepConfig } from '../steps/step-4-sustainability-practices/schema'
-
-// Import sustainability initiatives from config
-import { sustainabilityInitiativesStepConfig } from '@/lib/forms/configs/sustainability-initiatives-config'
 // Step configurations array
 
 export const stepConfigurations = [
