+++ apps/web/src/components/forms/multi-step/steps/step-2-business-model/index.ts
