+++ apps/web/src/components/forms/multi-step/steps/step-3-sustainability-initiatives/__tests__/sustainability-schemas.test.ts
@@ -4,7 +4,7 @@ import {
   sustainabilityInitiativeSchema,
   sustainabilityInitiativesSchema,
   sustainabilityInitiativesStepSchema,
-} from '../sustainability-schemas'
+} from '../schema'
 
 describe('Sustainability Schemas', () => {
   describe('initiativeTypeSchema', () => {
